mismi-core: Also retry IOExceptions
Found that IOExceptions were occasionally bubbling up to the top
level and killing the `s3 download` process. Since we are already
retrying on `HttpException` we can just add retrying on
`IOException`. Unfortunately, this is not sufficient to fix the
`s3 download` problem as we are now getting some processes that
seem to hang.